Mostly tools and documents for production use. Some miscellaneous changes. Java.inc file can be read off filesystem rather than request to Tomcat.

+++ java/src/com/ekkitab/search/TimerSnapshot.java
@@ -31,12 +31,12 @@ public class TimerSnapshot {
 		count++;
 		total += timeInMillis;
 		if (timeInMillis > max) {
-			max = (int)timeInMillis;
+			max = timeInMillis;
 		}
 		if (timeInMillis < min) {
-			min = (int)timeInMillis;
+			min = timeInMillis;
 		}
-		average = (int)total/count;
+		average = total/count;
 	}
 
 }
